more abstract mixin code
