+++ sdks/python/apache_beam/io/archive/zipfilesystemmixin.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-"""Archive system."""
+"""Archive file system mixin base."""
 
 # pytype: skip-file
 
@@ -23,12 +23,12 @@ from __future__ import absolute_import
 
 from future.utils import iteritems
 
-from apache_beam.io.aws import s3io
 from apache_beam.io.filesystem import BeamIOError
 from apache_beam.io.filesystem import CompressedFile
 from apache_beam.io.filesystem import CompressionTypes
 from apache_beam.io.filesystem import FileMetadata
 from apache_beam.io.filesystem import FileSystem
+from apache_beam.io.archive.archivefilesystemmixinbase import ArchiveFileSystemMixinBase
 from six import BytesIO
 
 from zipfile import ZipFile, ZipInfo
@@ -36,58 +36,25 @@ import datetime
 
 __all__ = ['ZipFileSystemMixin']
 
-class ArchiveFileSystemMixin(FileSystem):
-  """An mixin for `FileSystem` representing the contents of an archive file.
-  
-  Usage:
-
-  class CustomFileSystem(ZipFileSystemMixin, LocalFileSystem):
-    pass
-
-  """
-  def __init__(self, *args, **kwargs):
-    self.archive_path = kwargs.pop("archive_path")
-    super().__init__(*args, **kwargs)
-
-  @classmethod
-  def scheme(cls):
-    """URI scheme for the FileSystem
-    """
-    return "NOSCHEME"
-
-class ZipFileSystemMixin(ArchiveFileSystemMixin):
+class ZipFileSystemMixin(ArchiveFileSystemMixinBase):
   """An mixin for `FileSystem` representing the contents of .zip files.
   
   Usage:
 
   class CustomFileSystem(ZipFileSystemMixin, LocalFileSystem):
     pass
+  system = CustomFileSystem(pipeline_options=pipeline_options, archive_path=archive_path)
 
   """
   
   @classmethod
   def extension(cls):
-    """File extensions
+    """File extension
     """
     return "zip"
 
   def _read_archive_file(self):
-    return ZipFile(super().open(self.archive_path), "r")
-
-  def mkdirs(self, path):
-    """Recursively create directories for the provided path.
-
-    Args:
-      path: string path of the directory structure that should be created
-
-    Raises:
-      IOError: if leaf directory already exists.
-    """
-    pass
-
-  def has_dirs(self):
-    """Whether this FileSystem supports directories."""
-    return False
+    return ZipFile(super(ArchiveFileSystemMixinBase, self).open(self.archive_path), "r")
 
   def _list(self, dir_or_prefix):
     """List files in a location.
@@ -132,63 +99,6 @@ class ZipFileSystemMixin(ArchiveFileSystemMixin):
       if compression_type == CompressionTypes.UNCOMPRESSED:
         return raw_file
       return CompressedFile(raw_file, compression_type=compression_type)
-
-  def create(
-      self,
-      path,
-      mime_type='application/octet-stream',
-      compression_type=CompressionTypes.AUTO):
-    """Returns a write channel for the given file path.
-
-    Args:
-      path: string path of the file object to be written to the system
-      mime_type: MIME type to specify the type of content in the file object
-      compression_type: Type of compression to be used for this object
-
-    Returns: file handle with a close function for the user to use
-    """
-    return self._path_open_helper(path, 'wb', mime_type, compression_type)
-
-  def open(
-      self,
-      path,
-      mime_type='application/octet-stream',
-      compression_type=CompressionTypes.AUTO):
-    """Returns a read channel for the given file path.
-
-    Args:
-      path: string path of the file object to be written to the system
-      mime_type: MIME type to specify the type of content in the file object
-      compression_type: Type of compression to be used for this object
-
-    Returns: file handle with a close function for the user to use
-    """
-    return self._path_open_helper(path, 'rb', mime_type, compression_type)
-
-  def copy(self, source_file_names, destination_file_names):
-    """Recursively copy the file tree from the source to the destination
-
-    Args:
-      source_file_names: list of source file objects that needs to be copied
-      destination_file_names: list of destination of the new object
-
-    Raises:
-      ``BeamIOError``: if any of the copy operations fail
-    """
-    raise NotImplementedError
-
-  def rename(self, source_file_names, destination_file_names):
-    """Rename the files at the source list to the destination list.
-    Source and destination lists should be of the same size.
-
-    Args:
-      source_file_names: List of file paths that need to be moved
-      destination_file_names: List of destination_file_names for the files
-
-    Raises:
-      ``BeamIOError``: if any of the rename operations fail
-    """
-    raise NotImplementedError
 
   def exists(self, path):
     """Check if the provided path exists on the FileSystem.
@@ -263,11 +173,3 @@ class ZipFileSystemMixin(ArchiveFileSystemMixin):
     except Exception as e:  # pylint: disable=broad-except
       raise BeamIOError("Checksum operation failed", {path: e})
 
-  def delete(self, paths):
-    """Deletes files or directories at the provided paths.
-    Directories will be deleted recursively.
-
-    Args:
-      paths: list of paths that give the file objects to be deleted
-    """
-    raise NotImplementedError
